pyproject.toml
[project]
name = "testcase-generator"
version = "0.1.0"
description = "A comprehensive tool for generating test cases from requirements, prototypes, and screenshots"
readme = "README.md"
authors = [
    { name = "DamnPatrick", email = "patrickwuke@163.com" }
]
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "typer>=0.9.0",
    # "pydantic>=2.10.0",  # Temporarily disabled due to Python 3.14 compatibility
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
    "pandas>=2.0.0",
    
    # LLM integration - temporarily disabled due to Python 3.14 compatibility
    # "openai>=1.0.0",
    # "anthropic>=0.3.0",
    
    # Parsing dependencies
    "beautifulsoup4>=4.12.0",
    "python-docx>=0.8.11",
    "PyPDF2>=3.0.0",
    # "paddleocr>=2.7.0",  # Temporarily disabled due to Python 3.14 compatibility
    "opencv-python>=4.8.0",
    # "tesseract>=0.1.3",  # Temporarily disabled due to Python 3.14 compatibility
    
    # Web dependencies (for future UI)
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "jinja2>=3.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.0",
    
    # Testing and development
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[project.scripts]
testcase-gen = "testcase_generator.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["testcase_generator"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=testcase_generator",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
